requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.1.0
Werkzeug==2.3.7
pyrebase4>=4.7.1
firebase-admin==5.0.3
blinker>=1.6.2
jinja2>=3.1.2